Set expired TGT to empty string in local storage
Setting the TGT value in local storage to null
results in a read of a string containing the text "null"
and not the value null.

This results in a truthy value causes the guard checks
around TGT processing to get the string "null"
instead of the TGT value.

Since empty string is falsey, opting to set the expired
TGT to "" results in the guard check being skipped.