Extend okta's `isAuthenticated` check to require an active okta session
`isAuthenticated` is Okta's recommended flag for checking if a user is
logged in. By default it only verifies that both an idToken and accessToken
are available in the `TokenManager`. It does not verify if the tokens are
still alive (aka haven't been revoked).

As a result, the tokens can be saved off prior to sign out and re-used
at-will.

Verifying token status can be accomplished several ways, as there
are multiple endpoints that will fail out should the tokens be invalid.

Checking the Okta session is a good candidate since it is created and
destroyed by the signIn and signOut events respectively.

In fact, calling `oktaAuth.signOut()` signs the user of their current Okta session,
and, by default, revokes the accessToken and clears out the `TokenManager`,
indicating both session and token should be dealt with when logging users out
of the application anyways.

To add any additional check to `isAuthenticated`, pass a customized
`transformedAuthState` function to the OktaAuth config, which will
modify the `authState` object before it is stored and emitted.

This additional check will set `isAuthenticated` to the result
of  `session.exists()`. `False` indicates no active session, `True`,
an active exists.